Speed up per-frame hot paths and offline CSV pipeline
Across-the-board micro-optimisations driven by the bench suite —
no behaviour change, all 135 tests pass, savgol output is verified
byte-identical to the previous pandas path.

Highlights:
- match_hands_to_arms: stack-once + np.hypot replaces per-element
  list-comp + np.linalg.norm  (-25% to -44%).
- select_primary_body: 3-way size dispatch (scalar / loop / vectorised)
  recovers from the n=2 ufunc-dispatch regression  (-21% to -93%).
- _synthesise_hand_detections / _recrop_from_landmarks: scalar overlap
  loops beat numpy at the 1-6 element sizes we see in practice
  (-50% to -68% with palm_dets > 0).
- _smooth_detections: cost matrix via np.hypot on stacked centres
  (-17% to -38%).
- BoneLengthSmoother.update: single fancy-index + einsum replaces 10
  np.linalg.norm calls; in-place EMA; vectorised violation mask.
- OneEuroFilter: scalar-confidence fast path, aliased state copies,
  cached derivative-filter time constant.
- drawing: pre-blended limb-colour lookup tables, tolist() once per
  body for cv2 dot/line index reads, float32 throughout the spline.
- MetricsCollector.record: f-string formatting feeds csv.writer
  pre-stringified fields (skips the internal str(float) step),
  jitter computation inlined  (-17% to -21% per-frame).
- savgol_smooth_csv: bypass pandas to_csv's per-cell notna+format
  loop, keep smoothed data in a standalone numpy block (no df.loc
  write-back), skip pd.to_numeric on already-numeric columns
  (-54% to -69%).
- Bench fixtures use float32 to match the OpenVINO landmark dtype
  the production pipeline sees.

+++ scripts/benchmarks/bench_drawing.py
@@ -26,7 +26,9 @@ from ._harness import run_group
 
 def _spline_case(n_points: int, samples: int):
     rng = np.random.default_rng(5000 + n_points)
-    pts = rng.uniform(100, 600, (n_points, 2))
+    # Production callers pass float32 landmark slices; matching that here
+    # avoids a hidden float64→float32 copy at the start of every call.
+    pts = rng.uniform(100, 600, (n_points, 2)).astype(np.float32)
 
     def _run():
         catmull_rom_spline(pts, num_samples=samples)
